fix(STA-101): update integration tests for auth-aware DTOs
- Remove $.userId assertions from WalletApiIntegrationTest and E2E test
- Remove senderId from request body and response assertions in E2E test
- Replace jwt() with authentication(authenticationFor(userId)) so
  @AuthenticationPrincipal resolves to AuthPrincipal correctly
- Move authenticationFor helper to AuthFixtures (testFixtures) for reuse
- Add spring-security-oauth2-jose/resource-server to testFixtures deps
