+++ backend/src/test/java/com/stablepay/testutil/TestSecurityConfig.java
@@ -1,6 +1,5 @@
 package com.stablepay.testutil;
 
-import java.util.Collections;
 import java.util.UUID;
 
 import org.springframework.boot.test.context.TestConfiguration;
@@ -8,11 +7,7 @@ import org.springframework.context.annotation.Bean;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
 import org.springframework.security.core.Authentication;
-import org.springframework.security.oauth2.jwt.Jwt;
-import org.springframework.security.oauth2.server.resource.authentication.JwtAuthenticationToken;
 import org.springframework.security.web.SecurityFilterChain;
-
-import com.stablepay.domain.auth.model.AuthPrincipal;
 
 @TestConfiguration
 @EnableWebSecurity
@@ -26,16 +21,6 @@ public class TestSecurityConfig {
     }
 
     public static Authentication authenticationFor(UUID userId) {
-        var jwt = Jwt.withTokenValue("test-token")
-                .header("alg", "none")
-                .subject(userId.toString())
-                .build();
-        var principal = AuthPrincipal.builder().id(userId).build();
-        return new JwtAuthenticationToken(jwt, Collections.emptyList(), principal.id().toString()) {
-            @Override
-            public Object getPrincipal() {
-                return principal;
-            }
-        };
+        return AuthFixtures.authenticationFor(userId);
     }
 }
